Order output according to project name list
Added an optional parameter for supplying a list of projects. The
output will have the same order as this list. This makes it easier to
just copy paste all projects in one go. Added such a file with the
current order in the target document.

Also add dates on output rather than storing them in the data.